requirements.txt
Django==5.2
django-environ==0.12.0
psycopg2-binary==2.9.10
django-cors-headers==4.7.0
django-filter==25.1
djangorestframework==3.16.0
djangorestframework-simplejwt==5.5.0
gunicorn==23.0.0
ipython==9.4.0
drf-yasg==1.21.10
whitenoise==6.9.0
watchdog==6.0.0
django_celery_beat==2.8.1
redis==6.2.0
Celery==5.5.3
pandas==2.3.1